requirements.txt
duckduckgo-search
streamlit
smolagents
litellm
langchain
langchain-community
langchain-chroma
langchain-google-genai
langchain-ollama
chromadb
pdfplumber
groq
google-generativeai
pandas
numpy
beautifulsoup4
requests
matplotlib
openpyxl
